% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiple two numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
The product of two numbers
}
\description{
Multiple two numbers
}
